tier2/requirements.txt
flask
mysql-connector-python
opentelemetry-distro
opentelemetry-exporter-otlp
